# backend/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CollabMate API",
    "version": "1.0.0",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/team": {
      "post": {
        "summary": "Create a new team",
        "tags": [
          "Team"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "teamName"
                ],
                "properties": {
                  "teamName": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Team created"
          }
        }
      }
    },
    "/team/{teamId}/member": {
      "patch": {
        "summary": "Add a user to a team",
        "tags": [
          "Team"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "teamId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userId"
                ],
                "properties": {
                  "userId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Member added to team"
          }
        }
      }
    },
    "/team/{teamId}/role": {
      "patch": {
        "summary": "Assign a role to a team member",
        "tags": [
          "Team"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "teamId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userId",
                  "role"
                ],
                "properties": {
                  "userId": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Role assigned"
          }
        }
      }
    },
    "/team/{teamId}": {
      "get": {
        "summary": "Get team information by ID",
        "tags": [
          "Team"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "teamId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Team object returned"
          }
        }
      },
      "delete": {
        "summary": "Delete a team by ID (cascade clear user.teamId)",
        "description": "Deletes a team by its ID.\nAlso clears the `teamId` field of all users who belonged to this team.\n",
        "tags": [
          "Team"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "teamId",
            "required": true,
            "description": "The ID of the team to delete",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Team deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "team": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Team not found"
          }
        }
      }
    },
    "/team/{teamId}/members/detailed": {
      "get": {
        "summary": "Get detailed information of all team members",
        "description": "Returns detailed information for all members in a team, including userId, username, image, phone, role, status, and assigned tasks (taskId, title, status, deadline).\n",
        "tags": [
          "Team"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "teamId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Array of member detailed info",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "userId": {
                        "type": "string"
                      },
                      "username": {
                        "type": "string"
                      },
                      "image": {
                        "type": "string",
                        "nullable": true
                      },
                      "phone": {
                        "type": "string"
                      },
                      "role": {
                        "type": "string"
                      },
                      "status": {
                        "type": "string",
                        "enum": [
                          "online",
                          "busy",
                          "offline"
                        ]
                      },
                      "tasks": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "taskId": {
                              "type": "string"
                            },
                            "title": {
                              "type": "string"
                            },
                            "status": {
                              "type": "string"
                            },
                            "deadline": {
                              "type": "string",
                              "format": "date-time"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Team not found"
          }
        }
      }
    },
    "/team/{teamId}/leave": {
      "post": {
        "summary": "Let a user leave the team",
        "tags": [
          "Team"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "teamId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Team ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userId"
                ],
                "properties": {
                  "userId": {
                    "type": "string",
                    "description": "ID of the user leaving the team"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User left the team or team deleted"
          },
          "400": {
            "description": "User not in team"
          },
          "404": {
            "description": "Team not found"
          }
        }
      }
    },
...
  },
...
}